CORE: migrate BackCard and SharePost components to Deep Learning section
