+++ app/30dayschallenge/deep-learning/machine-learning/shareCard.jsx
